refactor(resolve-conflict): initiate ConflictResolver interface
