+++ pkg/conflictresolver/conflictresolver.go
@@ -1,0 +1,10 @@
+package conflictresolver
+
+type ConflictResolver interface {
+	Detect(fileNames []string) bool
+	Resolve(fileNames []string) error
+}
+
+var ConflictResolvers = []ConflictResolver{
+	&GoModResolver{},
+}
